Add automigrationの追加
